refactor: move llamalend deprecated markets and alerts in same file
