+++ apps/main/src/llamalend/features/market-list/hooks/useMarketAlert.tsx
@@ -1,58 +1,8 @@
-import { ReactNode, useMemo } from 'react'
+import { useMemo } from 'react'
+import { MARKETS_ALERTS } from '@/llamalend/llama-markets.constants'
 import { IChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import { Address } from '@primitives/address.utils'
-import { AlertType } from '@ui/AlertBox/types'
-import type { TooltipProps } from '@ui/Tooltip/types'
-import { t } from '@ui-kit/lib/i18n'
-import type { BannerProps } from '@ui-kit/shared/ui/Banner'
 import { LlamaMarketType } from '@ui-kit/types/market'
-import { Chain } from '@ui-kit/utils'
-
-export type MarketAlert = TooltipProps & {
-  alertType: AlertType
-  isDisableBorrow?: boolean // disallow user from creating new borrow positions
-  isDisableDeposit?: boolean // disallow user from supply deposit
-  // banner message, related to the market situation
-  banner?: Omit<BannerProps, 'children'> & { title: ReactNode }
-  // action card message, related to action of user
-  message?: ReactNode
-}
-
-export const MARKETS_ALERTS: Record<
-  LlamaMarketType,
-  { [chainId: number]: { [controllerAddress: Address]: MarketAlert } }
-> = {
-  /** LEND MARKET ALERTS */
-  Lend: {
-    [Chain.Ethereum]: {
-      // one-way-market-30 - sDOLA/crvUSD
-      '0xad444663c6c92b497225c6ce65fee2e7f78bfb86': {
-        alertType: 'danger',
-        isDisableBorrow: true,
-        isDisableDeposit: true,
-        message: t`This market is deprecated after a donation attack. New borrow positions and deposits are disabled.`,
-      },
-      // one-way-market-3 - CRV/crvUSD
-      '0xeda215b7666936ded834f76f3fbc6f323295110a': {
-        alertType: 'danger',
-        isDisableBorrow: true,
-        isDisableDeposit: true,
-        message: t`This market is deprecated. New borrow positions and deposits are disabled.`,
-      },
-    },
-    [Chain.Arbitrum]: {
-      // one-way-market-7 - FXN/crvUSD
-      '0x7adcc491f0b7f9bc12837b8f5edf0e580d176f1f': {
-        alertType: 'danger',
-        isDisableDeposit: true,
-        message: t`Due to small liquidity, borrowing or supplying in this market is not advisable.`,
-      },
-    },
-  },
-
-  /** MINT MARKET ALERTS */
-  Mint: {},
-}
 
 export const useMarketAlert = <ChainId extends IChainId>(
   rChainId: ChainId,
