+++ apps/main/src/llamalend/widgets/banners/DeprecatedMarketBanner.tsx
@@ -1,8 +1,8 @@
-import { DeprecatedMarket } from '@/llamalend/queries/market-list/constants'
+import { DeprecatedMarketAlert } from '@/llamalend/llama-markets.constants'
 import { t } from '@ui-kit/lib/i18n'
 import { Banner } from '@ui-kit/shared/ui/Banner'
 
-export const DeprecatedMarketBanner = ({ message, url }: DeprecatedMarket) => (
+export const DeprecatedMarketBanner = ({ message, url }: DeprecatedMarketAlert) => (
   <Banner severity="warning" subtitle={message} learnMoreUrl={url}>
     {t`This market is deprecated`}
   </Banner>
